requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
graphene==3.4.1
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
idna==3.10
promise==2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
